+++ RELEASES.md
@@ -1,9 +1,11 @@
 # NautilusTrader 1.113.0 Beta - Release Notes
 
-
+**This release includes substantial breaking changes.**
 
 ## Breaking Changes
-- None
+- Rename `AmendOrder` to `UpdateOrder`.
+- Rename `OrderAmended` to `OrderUpdated`.
+- Rename `amend` and `amended` related methods to `update` and `updated`.
 
 ## Enhancements
 - Refined LiveLogger: Now runs on event loop with high-performance `Queue`.